feat: add global post process api
